Move EncoderToTLV into DataModel as it looks like a more generic place, maybe we end up re-using it
