+++ src/app/data-model/BUILD.gn
@@ -19,12 +19,13 @@ source_set("data-model") {
     "DecodableList.h",
     "Decode.h",
     "Encode.h",
+    "EncoderToTLV.h",
     "FabricScoped.h",
     "FabricScopedPreEncodedValue.cpp",
     "FabricScopedPreEncodedValue.h",
     "List.h",
-    "NullObject.h",
     "Nullable.h",
+    "NullObject.h",
     "PreEncodedValue.cpp",
     "PreEncodedValue.h",
     "WrappedStructEncoder.h",
